Add ability to select sheets by sheet number or slug name
Previous versions required the exact case sensitive sheet name to
be entered.  This often required the user to have to put quotes
around the name and was generally a pain.

Instead of zenbones --sheet [name], now use --sheets [slug-name or #]
this makes it easier when using command-line history.  For example,
use zenbones --sheets to list the sheet numbers & slug names and just
up-arrow, add the name or number at the end of the command and presto,
it gets the sheet.